Cleaned up file names and added Archives directory for old versions of code
